dsds

+++ DATAWH/WebApplication1/WebApplication1/Datamodel/MongoMeasurement.cs
@@ -6,10 +6,10 @@ namespace WebApplication1.Datamodel
     public class MongoMeasurement
     {
         public ObjectId _id { get; set; }
-        public int light { get; set; }
+        public decimal light { get; set; }
         public decimal temperature { get; set; }
         public decimal humidity { get; set; }
-        public decimal co2 { get; set; }
+        public int co2 { get; set; }
         public int room_no { get; set; }
         public DateTime time { get; set; }
     }
